Hopefully, Jaben will not be too offended =) Else users would not use latest versions for a some indefinable time span as this all is not still completed. New changes need testing. (Added flags from ulong to int undoing to avoid errors.) Now YAF should work fine again.

+++ yafsrc/YAF.Classes/YAF.Classes.Data/Flags/ForumFlags.cs
@@ -119,7 +119,7 @@ namespace YAF.Classes.Data
     /// Use for bit comparisons
     /// </summary>
     [Flags]
-    public enum Flags : ulong
+    public enum Flags : int
     {
       /// <summary>
       /// The is locked.
